# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText API
  version: "4.0"
  description: |
    API specification for the WASAText messaging platform.
servers:
  - url: http://localhost:3000
    description: |
      Local development server
...
paths: 
...
  /users:
    get:
      tags: ["users"]
      summary: Search for users
      description: |
        Allows a user to search for other users by username. Returns a list of matching usernames.
        If a blank search (empty string or just spaces) is used, it will return all users.
      operationId: searchUsers
      security:
        - UserIdentifierAuth: []
      parameters:
        - name: query
          in: query
          required: false
          description: |
            Partial or full username to search for. Use a blank search to retrieve all users.
          schema:
            type: string
            description: |
              Search query
            pattern: '^[a-zA-Z0-9_-]{0,16}$'
            minLength: 0
            maxLength: 16
            example: "Pat"
      responses:
        "200":
          description: |
            List of matching users retrieved successfully
          content:
            application/json:
              schema:
                type: object
                description: |
                  Search results containing matching users
                properties:
                  users:
                    type: array
                    description: |
                      List of matching users
                    minItems: 0
                    maxItems: 1000
                    items:
                      type: object
                      description: |
                        User information
                      properties:
                        username:
                          type: string
                          description: |
                            Username of the user
                          pattern: '^[a-zA-Z0-9_-]{3,16}$'
                          minLength: 3
                          maxLength: 16
                          example: "Patric"
                        userId:
                          type: string
                          description: |
                            Unique identifier of the user
                          pattern: '^[a-zA-Z0-9_-]{12}$'
                          minLength: 12
                          maxLength: 12
                          example: "user12758923"
                        profilePhotoId:
                          type: string
                          description: |
                            Identifier of the user's profile photo
                          pattern: "^[a-zA-Z0-9_-]{10,30}$"
                          minLength: 10
                          maxLength: 30
                          example: "photo_789012"
                  total:
                    type: integer
                    description: |
                      Total number of matching users
                    minimum: 0
                    maximum: 1000
                    example: 5
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
components:
  securitySchemes:
    UserIdentifierAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        User identifier sent by the server in the login.
...
  responses:
    Unauthorized:
      description: |
        The user identifier is missing or is expired
      content:
        application/json:
          schema:
            type: object
            description: |
              Error details for unauthorized access
            properties:
              error:
                type: string
                description: |
                  Error message
                example: "Unauthorized access"
                pattern: '^[a-zA-Z0-9_ ]{10,100}$'
                minLength: 10
                maxLength: 100
    BadRequest:
      description: |
        The request was not compliant with the documentation (eg. missing fields, etc)
      content:
        application/json:
          schema:
            type: object
            description: |
              Error details for bad request
            properties:
              error:
                type: string
                description: |
                  Error message
                example: "Invalid request parameters"
                pattern: '^[a-zA-Z0-9_ ]{10,100}$'
                minLength: 10
                maxLength: 100
    InternalServerError:
      description: |
        The server encountered an internal error. Further info in server logs
      content:
        application/json:
          schema:
            type: object
            description: |
              Error details for internal server error
            properties:
              error:
                type: string
                description: |
                  Error message
                example: "Internal server error occurred"
                pattern: '^[a-zA-Z0-9_ ]{10,100}$'
                minLength: 10
                maxLength: 100
...
security:
  - UserIdentifierAuth: []